requirements.txt
websockets>=12.0
msgspec>=0.18
pyarrow>=14.0
pyyaml>=6.0
structlog>=23.0
